+++ src/connection/BaseConnectionOptions.ts
@@ -70,7 +70,7 @@ export interface BaseConnectionOptions {
 
     /**
      * Indicates if database schema should be auto created on every application launch.
-     * Be careful with this option and don't use this in production - otherwise you can loose production data.
+     * Be careful with this option and don't use this in production - otherwise you can lose production data.
      * This option is useful during debug and development.
      * Alternative to it, you can use CLI and run schema:sync command.
      *
@@ -87,7 +87,7 @@ export interface BaseConnectionOptions {
 
     /**
      * Drops the schema each time connection is being established.
-     * Be careful with this option and don't use this in production - otherwise you'll loose all production data.
+     * Be careful with this option and don't use this in production - otherwise you'll lose all production data.
      * This option is useful during debug and development.
      */
     readonly dropSchema?: boolean;