+++ test/functional/database-schema/column-types/postgres/entity/Post.ts
@@ -236,10 +236,12 @@ export class Post {
     array: number[];
 
     // -------------------------------------------------------------------------
-    // TypeOrm Specific Type
+    // TypeOrm Specific Types
     // -------------------------------------------------------------------------
 
     @Column("simple-array")
     simpleArray: string[];
 
+    @Column("simple-json")
+    simpleJson: { param: string };
 }