+++ test/utils/test-setup.ts
@@ -1,0 +1,6 @@
+import "reflect-metadata";
+import * as chai from "chai";
+
+chai.should();
+chai.use(require("sinon-chai"));
+chai.use(require("chai-as-promised"));