+++ CHANGELOG.md
@@ -54,10 +54,19 @@ By default its true.
 * now array initializations for relations are forbidden and ORM throws an error if there are entities with initialized relation arrays.
 * `@ClosureEntity` decorator has been removed. Instead `@Entity` + `@Tree("closure-table")` must be used
 * added support for nested set and materialized path tree hierarchy patterns
+* breaking change on how array parameters work in queries - now instead of (:param) new syntax must be used (:...param).
+This fixed various issues on how real arrays must work
+
+## 0.1.13
+
+* added simple-json column type ([#1448](https://github.com/typeorm/typeorm/pull/1488))
+* fixed transform behaviour for timestamp columns ([#1140](https://github.com/typeorm/typeorm/issues/1140))
+* fixed issue with multi-level relations loading ([#1504](https://github.com/typeorm/typeorm/issues/1504))
 
 ## 0.1.12
 
 * EntitySubscriber now fires events on subclass entity ([#1369](https://github.com/typeorm/typeorm/issues/1369))
+* fixed error with entity schema validator being async  ([#1448](https://github.com/typeorm/typeorm/issues/1448))
 
 ## 0.1.11